🪲 [Fix]: Base64 and show files on host (#67)
## Description

This pull request introduces several new PowerShell functions for Base64
encoding and decoding, as well as a minor update to an existing file
content display function. The most important changes include the
addition of three new functions for Base64 operations and a modification
to the `Show-FileContent` function to use `Write-Host`.

### New Base64 Functions:

*
[`src/functions/public/Base64/ConvertFrom-Base64.ps1`](diffhunk://#diff-f4a025080ef7de3f1252eb0f6c962b8c4a7ae9436929eb6aea500c8357e650d6R1-R24):
Added a new function `ConvertFrom-Base64` to convert a Base64 encoded
string to a regular string.
*
[`src/functions/public/Base64/ConvertTo-Base64.ps1`](diffhunk://#diff-6bd8dc56f742df93ae23ac26e58a028c4ef6dc664f3d994a195f7069021995d2R1-R23):
Added a new function `ConvertTo-Base64` to convert a regular string to a
Base64 encoded string.
*
[`src/functions/public/Base64/Test-Base64.ps1`](diffhunk://#diff-2cf7c7018923286250348248fa89bf0811869774e95aec4814296090e1242c48R1-R26):
Added a new function `Test-Base64` to check if a string is a valid
Base64 encoded string.

### Update to Existing Function:

*
[`src/functions/public/Files/Show-FileContent.ps1`](diffhunk://#diff-123988de399ad37f4694e6a1fa20fb3f8cee97a12400398f0ba87ef30bb6cab4L29-R29):
Modified the `Show-FileContent` function to use `Write-Host` for
outputting the formatted line number and content.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas