🚀 [Feature]: Moved `Guid` functions to the `Guid` module (#72)
## Description

This pull request involves the removal of two PowerShell filters and
their corresponding tests. The filters `Search-GUID` and `Test-IsGUID`
were used to search for and test GUIDs within strings. The changes
include the deletion of these filters and their associated test cases.

These function have moved to [`GUID`](https://github.com/PSModule/GUID).

Removal of filters:

*
[`src/functions/public/GUID/Search-GUID.ps1`](diffhunk://#diff-cc702758c32f07e846a3eb171fe5887d819f4bae0d3fa31023fcaa0ffc1e8f4cL1-L31):
Removed the `Search-GUID` filter, which was used to search a string for
a GUID.
*
[`src/functions/public/GUID/Test-IsGUID.ps1`](diffhunk://#diff-924a7016be2c382a50a56c882ebc4201befae5abe6feb805cdb9a9c38b5ff81fL1-L31):
Removed the `Test-IsGUID` filter, which was used to test if a string is
a GUID.

Removal of tests:

*
[`tests/Utilities.Tests.ps1`](diffhunk://#diff-c47a3a4aeb91081376e3f1332962b18b0f21f480d0e77fbac0b6b7a045625e38L183-L196):
Deleted test cases for `Search-GUID` and `Test-IsGUID` filters.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [x] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas