🚀 [Feature]: Moved `Casing` functions to the `Casing` module (#73)
## Description

This pull request involves the removal of several PowerShell functions
related to string casing and their associated tests. The most important
changes include the deletion of `Convert-StringCasingStyle`,
`Get-StringCasingStyle`, and `Split-StringByCasingStyle` functions,
along with the removal of their corresponding test cases.

These function have moved to
[Casing](https://github.com/PSModule/Casing).

### Removal of string casing functions:

*
[`src/functions/public/String/Casing/Convert-StringCasingStyle.ps1`](diffhunk://#diff-3c0f07d3c04fe8a1b04c62890e1970fa6e14229f57ecd7536b172cd107539f48L1-L71):
Deleted the entire `Convert-StringCasingStyle` function, which was
responsible for converting a string to different casing styles.
*
[`src/functions/public/String/Casing/Get-StringCasingStyle.ps1`](diffhunk://#diff-df27d8bdb3dc074f4d357d904f5ad1c974117150a0ca75275f827fdb604d817eL1-L98):
Deleted the entire `Get-StringCasingStyle` function, which was used to
detect the casing style of a string.
*
[`src/functions/public/String/Casing/Split-StringByCasingStyle.ps1`](diffhunk://#diff-8320615a3057764e1493d3da5305b3573a0f475d7efea8528755276a3cedc45eL1-L101):
Deleted the entire `Split-StringByCasingStyle` function, which split
strings into arrays of words based on their casing style.

### Removal of test cases:

*
[`tests/Utilities.Tests.ps1`](diffhunk://#diff-c47a3a4aeb91081376e3f1332962b18b0f21f480d0e77fbac0b6b7a045625e38L183-L268):
Removed all test cases related to the `Get-StringCasingStyle` function,
which tested various string casing styles.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [x] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas