🚀 [Feature]: Move `TLS` to the `TLS` module (#80)
## Description

This pull request removes `Get-TLSConfig` and `Set-TLSConfig`, which
means that the capabilities to retrieve and set the TLS configuration in
the current PowerShell session are no longer available here.

These functions have moved to the
[`TLS`](https://github.com/PSModule/TLS/) module.

Removal of TLS configuration functions:

*
[`src/functions/public/TLS/Get-TLSConfig.ps1`](diffhunk://#diff-be7457a199f41686bf47826defb9350111d8f18318b21896be359c3732eb15ebL1-L31):
Deleted the function `Get-TLSConfig` which was used to get the TLS
configuration of the current session.
*
[`src/functions/public/TLS/Set-TLSConfig.ps1`](diffhunk://#diff-491b4928d15474e12b8f82638799f1636e195d450f70d78ab944b6b6bc951593L1-L28):
Deleted the function `Set-TLSConfig` which was used to set the TLS
configuration for the current PowerShell session.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [x] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas