🚀 [Feature]: Move Hashtable functions to `Hashtable` module (#82)
## Description


This pull request involves the removal of several functions and their
associated tests related to hashtable operations. The changes include
the deletion of `Convert-HashtableToString`, `ConvertTo-Hashtable`, and
`Merge-Hashtable` functions, as well as the removal of their
corresponding test cases.

The hashtable functions are now moved to the
**[`Hashtable`](https://github.com/PSModule/Hashtable)** module.

Removed functions and their tests:

*
[`src/functions/public/Hashtable/Convert-HashtableToString.ps1`](diffhunk://#diff-60fc15aad7d10c312fe92888c7671d146b16b1d744d2edb739fdbbd55d0a1135L1-L103):
Deleted the `Convert-HashtableToString` function which was responsible
for converting a hashtable to its code representation.
*
[`src/functions/public/Hashtable/ConvertTo-Hashtable.ps1`](diffhunk://#diff-c51782aa189f0ed6ecfa4c4030f596b5cd4d73e798b2be5cf435295f357bf436L1-L70):
Deleted the `ConvertTo-Hashtable` function which converted an object to
a hashtable.
*
[`src/functions/public/Hashtable/Merge-Hashtable.ps1`](diffhunk://#diff-288b85c7499e29e910cb141e6b1b65a598ef80a7068827e034118cfefa6de623L1-L56):
Deleted the `Merge-Hashtable` function which merged two hashtables, with
the second hashtable overriding the first.
*
[`tests/Utilities.Tests.ps1`](diffhunk://#diff-c47a3a4aeb91081376e3f1332962b18b0f21f480d0e77fbac0b6b7a045625e38L119-L326):
Removed test cases for the `Merge-Hashtable` and `ConvertTo-Hashtable`
functions.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [x] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas